+++ src/main/java/org/esup_portail/esup_stage/controller/PeriodeInterruptionStageController.java
@@ -7,19 +7,24 @@ import org.esup_portail.esup_stage.dto.view.Views;
 import org.esup_portail.esup_stage.enums.AppFonctionEnum;
 import org.esup_portail.esup_stage.enums.DroitEnum;
 import org.esup_portail.esup_stage.exception.AppException;
-import org.esup_portail.esup_stage.model.Convention;
-import org.esup_portail.esup_stage.model.PeriodeInterruptionStage;
+import org.esup_portail.esup_stage.model.*;
+import org.esup_portail.esup_stage.model.helper.UtilisateurHelper;
 import org.esup_portail.esup_stage.repository.ConventionJpaRepository;
 import org.esup_portail.esup_stage.repository.PeriodeInterruptionStageJpaRepository;
 import org.esup_portail.esup_stage.repository.PeriodeInterruptionStageRepository;
+import org.esup_portail.esup_stage.security.ServiceContext;
 import org.esup_portail.esup_stage.security.interceptor.Secure;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @ApiController
 @RequestMapping("/periode-interruption-stage")
@@ -38,7 +43,16 @@ public class PeriodeInterruptionStageController {
     @GetMapping
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.LECTURE})
     public PaginatedResponse<PeriodeInterruptionStage> search(@RequestParam(name = "page", defaultValue = "1") int page, @RequestParam(name = "perPage", defaultValue = "50") int perPage, @RequestParam("predicate") String predicate, @RequestParam(name = "sortOrder", defaultValue = "asc") String sortOrder, @RequestParam(name = "filters", defaultValue = "{}") String filters, HttpServletResponse response) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
         PaginatedResponse<PeriodeInterruptionStage> paginatedResponse = new PaginatedResponse<>();
+        if (UtilisateurHelper.isRole(utilisateur, Role.ETU)) {
+            JSONObject jsonFilters = new JSONObject(filters);
+            Map<String, Object> currentUser = new HashMap<>();
+            currentUser.put("type", "text");
+            currentUser.put("value", utilisateur.getLogin());
+            jsonFilters.put("convention.etudiant.identEtudiant", currentUser);
+            filters = jsonFilters.toString();
+        }
         paginatedResponse.setTotal(periodeInterruptionStageRepository.count(filters));
         paginatedResponse.setData(periodeInterruptionStageRepository.findPaginated(page, perPage, predicate, sortOrder, filters));
         return paginatedResponse;
@@ -48,8 +62,9 @@ public class PeriodeInterruptionStageController {
     @GetMapping("/{id}")
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.LECTURE})
     public PeriodeInterruptionStage getById(@PathVariable("id") int id) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
         PeriodeInterruptionStage periodeInterruptionStage = periodeInterruptionStageJpaRepository.findById(id);
-        if (periodeInterruptionStage == null) {
+        if (periodeInterruptionStage == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(periodeInterruptionStage.getConvention().getEtudiant().getIdentEtudiant()))) {
             throw new AppException(HttpStatus.NOT_FOUND, "PeriodeInterruptionStage non trouvée");
         }
         return periodeInterruptionStage;
@@ -59,12 +74,22 @@ public class PeriodeInterruptionStageController {
     @GetMapping("/getByConvention/{id}")
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.LECTURE})
     public List<PeriodeInterruptionStage> getByConvention(@PathVariable("id") int id) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
+        Convention convention = conventionJpaRepository.findById(id);
+        if (convention == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(convention.getEtudiant().getIdentEtudiant()))) {
+            return new ArrayList<>();
+        }
         return periodeInterruptionStageJpaRepository.findByConvention(id);
     }
 
     @PostMapping
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.CREATION})
     public PeriodeInterruptionStage create(@Valid @RequestBody PeriodeInterruptionStageDto periodeInterruptionStageDto) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
+        Convention convention = conventionJpaRepository.findById(periodeInterruptionStageDto.getIdConvention());
+        if (convention == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(convention.getEtudiant().getIdentEtudiant()))) {
+            throw new AppException(HttpStatus.NOT_FOUND, "Convention non trouvéz");
+        }
         PeriodeInterruptionStage periodeInterruptionStage = new PeriodeInterruptionStage();
         setPeriodeInterruptionStageData(periodeInterruptionStage, periodeInterruptionStageDto);
         return periodeInterruptionStageJpaRepository.saveAndFlush(periodeInterruptionStage);
@@ -73,8 +98,9 @@ public class PeriodeInterruptionStageController {
     @PutMapping("/{id}")
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.MODIFICATION})
     public PeriodeInterruptionStage update(@PathVariable("id") int id, @Valid @RequestBody PeriodeInterruptionStageDto periodeInterruptionStageDto) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
         PeriodeInterruptionStage periodeInterruptionStage = periodeInterruptionStageJpaRepository.findById(id);
-        if (periodeInterruptionStage == null) {
+        if (periodeInterruptionStage == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(periodeInterruptionStage.getConvention().getEtudiant().getIdentEtudiant()))) {
             throw new AppException(HttpStatus.NOT_FOUND, "PeriodeInterruptionStage non trouvée");
         }
         setPeriodeInterruptionStageData(periodeInterruptionStage, periodeInterruptionStageDto);
@@ -84,8 +110,9 @@ public class PeriodeInterruptionStageController {
     @DeleteMapping("/{id}")
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.SUPPRESSION})
     public boolean delete(@PathVariable("id") int id) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
         PeriodeInterruptionStage periodeInterruptionStage = periodeInterruptionStageJpaRepository.findById(id);
-        if (periodeInterruptionStage == null) {
+        if (periodeInterruptionStage == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(periodeInterruptionStage.getConvention().getEtudiant().getIdentEtudiant()))) {
             throw new AppException(HttpStatus.NOT_FOUND, "PeriodeInterruptionStage non trouvé");
         }
         periodeInterruptionStageJpaRepository.delete(periodeInterruptionStage);
@@ -96,6 +123,11 @@ public class PeriodeInterruptionStageController {
     @DeleteMapping("/delete-by-convention/{idConvention}")
     @Secure(fonctions = {AppFonctionEnum.CONVENTION}, droits = {DroitEnum.SUPPRESSION})
     public boolean deleteByConvention(@PathVariable("idConvention") int idConvention) {
+        Utilisateur utilisateur = ServiceContext.getServiceContext().getUtilisateur();
+        Convention convention = conventionJpaRepository.findById(idConvention);
+        if (convention == null || (UtilisateurHelper.isRole(utilisateur, Role.ETU) && !utilisateur.getLogin().equals(convention.getEtudiant().getIdentEtudiant()))) {
+            throw new AppException(HttpStatus.NOT_FOUND, "Convention non trouvée");
+        }
         periodeInterruptionStageJpaRepository.deleteByConvention(idConvention);
         periodeInterruptionStageJpaRepository.flush();
         return true;
